Cleaned pipex and executor Fixed #148
