+++ Makefile
@@ -1,14 +1,11 @@
 CC			= gcc
-#CFLAGS		= -g3
-# CFLAGS		= -g3 -Wall -Wextra -Werror
-CFLAGS		=
-#CFLAGS		=
+CFLAGS		= -g3 -Wall -Wextra -Werror
 RM			= rm -f
 SRCD		= ./srcs/
 SRC			= main.c ft_lexor.c ft_parser.c ft_executor.c ft_initiator.c ft_echo_utils.c ft_builtins2.c ft_builtins.c ft_globlal_fun.c \
 				ft_builtins_utils.c ft_cd_utils.c ft_lexor_utils.c ft_lexor_utils2.c ft_lexor_handlers.c ft_parser_utils.c utils.c \
 				ft_free_linked_list.c ft_exit_utils.c ft_export_utils.c get_next_line.c get_next_line_utils.c ft_init_utils.c \
-				pipex.c helper.c test.c exit_handler.c
+				pipex.c helper.c exit_handler.c
 
 # Command to add the source folder prefix (instead of having it added manually to SRC)
 SRCF		= $(addprefix $(SRCD),$(SRC))
